requirements/base.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
aws-cdk.core==1.17.1
aws-cdk.cx-api==1.17.1
awscli==1.18.8
AWSIoTPythonSDK==1.4.7
boto3==1.12.8
botocore==1.15.8
bottle==0.12.13
cattrs==0.9.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.6
django-braces==1.14.0
django-cors-headers==3.3.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
filelock==3.0.12
idna==2.10
importlib-metadata==1.6.1
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.4
jsii==0.20.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==2.13.5
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
packaging==20.4
pandas==1.0.4
peewee==2.10.1
Pillow==7.1.2
publication==0.0.3
pyasn1==0.4.7
pycparser==2.20
PyJWT==1.5.2
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.1.2
requests==2.24.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.12.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.6
virtualenv==20.0.24
wrapt==1.12.1
zipp==3.1.0
